requirements.txt
typing_extensions==4.12.2
anthropic==0.49.0
google-genai==1.9.0
